# api.yaml
swagger: "2.0"
info:
  title: Viviendas reales
  description: descripción de la API.
  version: 1.0.0
host: localhost:5000
basePath: /api
schemes:
  - http
  - https
paths:
  /user:
    post:
      tags:
        - user
      description: crea un nuevo usuario
      parameters:
        - name: user
          description: datos que representan al nuevo usuario
          in: body
          schema:
            type: object
            properties:
              name:
                type: string
              surname:
                type: string
              email:
                type: string
              residence:
                type: string
              password:
                type: string
              avatar:
                type: string
              role:
                type: string
                enum: [ user, admin ]
            required:
              - name
              - surname
              - email
              - password
      responses:
        201:
          description: Created
        400:
          description: Bad request
    get:
      tags:
        - user
      description: devuelve un único usuario
      parameters:
        - name: page
          description: página a la que accede la petición
          in: query
          required: false
          type: string
        - name: limit
          description: cantidad de recursos máximos devueltos
          in: query
          required: false
          type: string
      responses:
        200:
          description: Success
        400:
          description: Bad request
  /user/{id}:
    get:
      tags:
        - user
      description: Get all Employee
      parameters:
        - name: id
          description: Create an new employee
          in: path
          required: true
          type: string
      responses:
        200:
          description: Success
        400:
          description: Bad request
    put:
      security:
        - BearerAuth: [token]
      tags:
        - user
      description: Actualiza un usuario cuya id coincide con la pasada po parámetro
      parameters:
        - name: id
          description: ObjectID del usuario a borrar
          in: path
          required: true
          type: string
        - name: user
          in: body
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
              surname:
                type: string
              password:
                type: string
              email:
                type: string
              residence:
                type: string
              role:
                type: string
      responses:
        200:
          description: Success
        400:
          description: Bad request
    patch:
      tags:
        - user
      description: actualiza el avatar del usuario cuya id coincide con la pasada po parámetro
      parameters:
        - name: id
          description: ObjectID del usuario a borrar
          in: path
          required: true
          type: string
        - name: token
          description: token que autentica al usuario que realiza la petición
          in: header
          required: true
          type: string
      responses:
        200:
          description: Success
        400:
          description: Bad request
...
  /user{id}:
    delete:
      tags:
        - user
      description: Borra un usuario cuya id coincide con la pasada po parámetro
      parameters:
        - name: id
          description: ObjectID del usuario a borrar
          in: path
          required: true
          type: string
      responses:
        200:
          description: Success
        400:
          description: Bad request
  /comment:
    post:
      tags:
        - comment
      description: crea nuevo comentario
      parameters:
        - name: comment
          description: objeto el cual contiene los datos del nuevo usuario
          in: body
          schema:
            type: object
            properties:
              author:
                type: string
              content:
                type: string
              title:
                type: string
              housing:
                type: string
            required:
              - author
              - content
      responses:
        200:
          description: Success
        400:
          description: Bad request
    get:
      tags:
        - comment
      description: recupera los usuarios
      responses:
        200:
          description: success
        404:
          description: si no se encuentra
        400:
          description: error en algún parámetro
      parameters:
        - name: page
          description: página a la que accede la petición
          in: query
          required: false
          type: string
        - name: limit
          in: query
          required: false
          type: string
  /comment/{id}:
    get:
      tags:
        - comment
      description: recupera los datos de un comentario el cual coincide con l id que se pasa por parámetro
      responses:
        200:
          description: success
      parameters:
        - name: id
          description: id del comentario al cual se pretende acceder
          type: string
          in: path
          required: true
    delete:
      tags:
        - comment
      parameters:
        - name: id
          type: string
          description: id del recurse el cual se pretende borrar
          in: path
          required: true
      responses:
        203:
          description: cuando se ha borrado
        404:
          description: si no existe el recurso
    patch:
      tags:
        - comment
      description: actualiza el recurso correspondiente a la id pasada por parámetro
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        201:
          description: si se actualiza
        400:
          description: en caso de error
        404:
          description: si no existe el recurso
  /comment/like/{id}:
    patch:
      tags:
        - comment
      description: crea nuevo comentario
      responses:
        201:
          description: si se actualiza el recurso
        400:
          description: si hay algún error en la consulta
        404:
          description: si el recurso al que se pretende acceder no existe
      parameters:
        - name: comment
          description: objeto el cual contiene los datos del nuevo usuario
          in: body
          schema:
            type: object
            properties:
              author:
                type: string
              content:
                type: string
              title:
                type: string
              housing:
                type: string
            required:
              - author
              - content
        - name: id
          in: path
          required: true
          type: string
  /comment/like/remove/{id}:
    patch:
      tags:
        - comment
      description: crea nuevo comentario
      responses:
        201:
          description: si se actualiza el recurso
        400:
          description: si hay algún error en la consulta
        404:
          description: si el recurso al que se pretende acceder no existe
      parameters:
        - name: comment
          description: objeto el cual contiene los datos del nuevo usuario
          in: body
          schema:
            type: object
            properties:
              author:
                type: string
              content:
                type: string
              title:
                type: string
              housing:
                type: string
            required:
              - author
              - content
        - name: id
          in: path
          type: string
          required: true
...